Revert "Fix scanForWantedComponents not ignoring attribute values of 0."
This patch reverts commit 2648d9297f25a0d1fa2837f020975a45d4e8a8b9 as a
workaround for the "banding" artifacts we were seeing in Qt. Angle
returns a list of supported graphic formats or configurations, sorting
it in a way that the first one should be the one that fits better the
requested format. In Qt we use the first thing we receive in the list.
In the current Angle version, however, a fix has changed the way in
which the list is sorted. In the old version the first element would be
a 32-bit graphic format, while now it's a 16-bit one, resulting in the
"banding" artifacts. The workaround reverts back to the previous sorting
behavior.