ANGLE: Disable multisampling to avoid crash in Qt applications
This patch adds a workaround for the crash that occurs with multisampling
enabled in some Qt applications (e.g., the 2dpainting example). The old
Angle release we were using lacked support for multisampling, so it was
ignored when requested. This current version seems to support it, but is
causing a crash in the D3D11 shaders, so this workaround disables that
support.