D3D: Initialize storage after generating mipmap images.
When TextureD3D generates mipmaps without a storage present, it
initializes all the images and then generates mipmaps on the CPU.
Attempt to generate a storage after initializing all the mip images
so that the mipmap generation can be done on the GPU.

Fix several edge cases where texture storage was not copied back
to images when redefining for mipmap generation. This resulted in
the newly generated mipmaps not using the correct data.

Bug: chromium:1256340
Change-Id: I76122d36385450d973bf00f7d07402a1f975492c
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3406643
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Jamie Madill <jmadill@chromium.org>